--- packages/app/src/app/pages/Sandbox/Editor/Workspace/CreateRepo/elements.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const Error = styled.div`
-  margin: 1rem;
-  color: ${({ theme }) => theme.red};
-  font-size: 0.875rem;
-`;
