+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/GitHub/GitHub.tsx
@@ -1,22 +1,26 @@
 import React from 'react';
-import { inject, observer } from 'mobx-react';
-
-import GithubIntegration from '../../../../../common/GithubIntegration';
+import { observer } from 'mobx-react-lite';
+import GithubIntegration from 'app/pages/common/GithubIntegration';
+import { useStore } from 'app/store';
 import WorkspaceItem from '../../WorkspaceItem';
-import Git from '../../Git';
-import CreateRepo from '../../CreateRepo';
 import { Description } from '../../elements';
+import { Git } from './Git';
+import { CreateRepo } from './CreateRepo';
 
-const GitHub = ({ store }) => {
-  const sandbox = store.editor.currentSandbox;
+export const GitHub = observer(() => {
+  const {
+    editor: {
+      currentSandbox: { originalGit },
+    },
+    user: {
+      integrations: { github },
+    },
+  } = useStore();
 
-  const hasOriginalGitSource = sandbox.originalGit;
-
-  return store.user.integrations.github ? ( // eslint-disable-line
-    hasOriginalGitSource ? (
+  return github ? ( // eslint-disable-line
+    originalGit ? (
       <>
         <Git />
-
         <WorkspaceItem title="Export to GitHub">
           <CreateRepo />
         </WorkspaceItem>
@@ -28,7 +32,7 @@ const GitHub = ({ store }) => {
       </>
     )
   ) : (
-    <div>
+    <>
       <Description margin={1} top={0}>
         You can create commits and open pull requests if you add GitHub to your
         integrations.
@@ -37,8 +41,6 @@ const GitHub = ({ store }) => {
       <div style={{ margin: '1rem' }}>
         <GithubIntegration small />
       </div>
-    </div>
+    </>
   );
-};
-
-export default inject('signals', 'store')(observer(GitHub));
+});
