+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/GitHub/index.ts
@@ -1,0 +1,1 @@
+export { GitHub } from './GitHub';
